Use UInt64 instead of Double for dates

+++ CloverSDK/Models/TransformTypes.swift
@@ -11,20 +11,20 @@ import ObjectMapper
 
 public class CLVDateTransform: TransformType {
   public typealias Object = NSDate
-  public typealias JSON = Double
+  public typealias JSON = UInt64
   
   public init() {}
   
   public func transformFromJSON(value: AnyObject?) -> NSDate? {
-    if let timeInt = value as? Double {
+    if let timeInt = value as? UInt64 {
       return NSDate(timeIntervalSince1970: NSTimeInterval(timeInt / 1000))
     }
     return nil
   }
   
-  public func transformToJSON(value: NSDate?) -> Double? {
+  public func transformToJSON(value: NSDate?) -> UInt64? {
     if let date = value {
-      return Double(date.timeIntervalSince1970 * 1000)
+      return UInt64(date.timeIntervalSince1970 * 1000)
     }
     return nil
   }
